Refactored file locations.
